requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
sqlalchemy==2.0.34
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
